🐛 overflow woes

+++ Exercises/1001-1500/1076-1100/1093MaxScoreAfterApplyingKOps/max_score_after_applying_k_operations.cpp
@@ -61,12 +61,12 @@ public:
       for (int n: nums) {
         pq.push(n);
       }
-      int res = 0;
+      long long res = 0;
       for (int i = 0; i < k; ++i) {
         int cur = pq.top();
         pq.pop();
         res+=cur;
-        float quo = (float)cur/3.0;
+        double quo = (double)cur/3.0;
         pq.push(ceil(quo));
       }
       return res;
